requirements.txt
langchain
langchain-community
langchain-core
langchain-ollama
ollama
requests
google-api-python-client
google-auth-httplib2
google-auth-oauthlib